Add sparse encoding prototype
This creates an optional sparse encoding scheme where unset registers
are not stored. Set registers are stored in a sorted linked list. A temporary
buffer is used for insertions into this list. This buffer is
periodically cleared (when it is full or cardinality() is called) by
first sorting it and then trying to insert each item into the linked
list of registers.