Insert all nodes in the buffer using only one pass
Instead of re-traversing the linked list of registers from start to
finish for each node we want to insert, we instead keep a pointer to the
last updated node in the list. This allows us to traverse the linked
list only once since we know for each node in the buffer, no subsequent
nodes will have a greater index so there is no need to inspect nodes in
the linked list before the last updated node.